requirements.txt
pyaudio==0.2.11
ffmpeg-python==0.2.0
pydrive2==1.18.1